+++ R/plot2rasters.R
@@ -167,7 +167,7 @@ setMethod("plot", signature(x="SpatRaster", y="SpatRaster"),
 		if (is.null(xlim)) xlim <- c(xmin(x), xmax(x))
 		if (is.null(ylim)) ylim <- c(ymin(x), ymax(x))
 		e <- extent(xlim, ylim)
-		out <- expand(crop(out, e), e, value=0)
+		out <- extend(crop(out, e), e, value=0)
 	}
 	plot(out, maxcell=nc*nr, asp=asp, ...) 
 }